move wolfios to its own package
